Make a lot of improvements and fix bugs
Fully remade style setup of Header menus, Popup menu and LeftSideMenu so
they're using better algorithm and they automatically recalculate styles
on resize, so they don't close on resize which brings better UX.
I've fixed bug with scroll on IOS when modal windows are open.
I've fixed bug with interactive items in HeaderMainMenu when it's
closed.
Now Select element closes when it's open and user clicks outside select.
When user follows the link page automatically scrolls to top.
I've fixed bug with Slider.
And now technical page GetAnalogueProducts which is used only to operate
with POST requests throws 404 error if somebody open it, because as a
real page it's actually doesn't exist.